requirements.txt
numpy==1.22.4
opencv-python==4.6.0.66
pyserial==3.5
